requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Cors>=3.0
python-dotenv>=0.20
PyMySQL>=1.0
Flask-JWT-Extended>=4.4
bcrypt>=4.0
marshmallow>=3.20